Separated `gen/immediate.go` into multiple files
